refactor(cfc): share conceptGuard from exchange-eval into atom-pattern
Move the concept-guard recognizer into atom-pattern.ts (the shared leaf
both consumers already import) and export it, so the exchange-rule
evaluator (B4 rule guards) and the integrity floors (D5, observation.ts)
recognize a Concept-typed entry identically. Pure move; no behavior change.
