+++ packages/runner/src/cfc/exchange-eval.ts
@@ -1,11 +1,9 @@
-import { CFC_ATOM_TYPE } from "@commonfabric/api/cfc";
 import { deepEqual } from "@commonfabric/utils/deep-equal";
-import { isRecord } from "@commonfabric/utils/types";
 import { utf8Compare } from "@commonfabric/utils/utf8";
 import {
   type AtomPatternBindings,
+  conceptGuard,
   instantiateAtomPattern,
-  isAtomVarPlaceholder,
   matchAtomPattern,
   matchAtomPatternAgainstAtoms,
 } from "./atom-pattern.ts";
@@ -83,44 +81,6 @@ export type ExchangeEvalResult = {
   readonly label: IFCLabel;
   readonly firings: readonly RuleFiring[];
   readonly exhausted: boolean;
-};
-
-/**
- * A concept-valued integrity guard (spec §4.4.5): a CONCRETE `Concept` atom
- * pattern. Satisfied exclusively via the acting principal's trust closure —
- * never by pool-matching a literal Concept atom (which carried integrity
- * cannot legitimately contain; the mint gate strips it). Returns `undefined`
- * for non-concept-shaped patterns (they route to ordinary pool matching) and
- * `{ uri: undefined }` — a concept guard that is NEVER satisfied — for any
- * `Concept`-typed pattern that is not the EXACT concrete two-field shape
- * `{ type, uri: <non-empty string> }`.
- *
- * The exact-shape requirement is load-bearing (codex/cubic P2 on #4564):
- * concept guards are checked ONLY on `type` + trust closure, so extra
- * constraint fields (`{ type: Concept, uri, subject: … }`) or a smuggled
- * `var` key would be silently ignored — an over-constrained guard the author
- * wrote as narrow would fire as broadly as the bare concept. Anything but the
- * canonical shape fails closed.
- */
-const conceptGuard = (
-  pattern: unknown,
-): { uri: string | undefined } | undefined => {
-  if (
-    !isRecord(pattern) || isAtomVarPlaceholder(pattern) ||
-    (pattern as { type?: unknown }).type !== CFC_ATOM_TYPE.Concept
-  ) {
-    return undefined;
-  }
-  const uri = (pattern as { uri?: unknown }).uri;
-  // Exactly `{ type, uri }`, uri a non-empty string. Extra fields (or a `var`
-  // key, which would push the key count past 2) → never satisfied.
-  if (
-    Object.keys(pattern).length !== 2 ||
-    typeof uri !== "string" || uri.length === 0
-  ) {
-    return { uri: undefined };
-  }
-  return { uri };
 };
 
 /**
